3-tier-dockerized-app-on-aws/private-subnet/backend-flask/requirements.txt
blinker
boto3==1.11.0
botocore==1.14.9
click==8.1.7
Flask
Flask-Cors
Flask-DotEnv
Flask-SQLAlchemy
greenlet==3.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath
MarkupSafe==2.1.3
psycopg2-binary==2.9.9
python-dateutil==2.8.2
python-dotenv
s3transfer
six==1.16.0
SQLAlchemy==2.0.23
typing_extensions
urllib3
Werkzeug
itsdangerous==2.1.2
